Move header parsing code to new module
