+++ autoendpoint/src/server/headers/util.rs
